+++ docs/concept/arrange.md
@@ -5,5 +5,17 @@
 ### 成员
 
 - 画布位置
+
+    布局中心的画布坐标系下的坐标。
+
 - 旋转
+
+    一个角度，顺时针方向。
+
 - 外接矩形
+
+    通过所属组件计算。计算来源于图形svg和文字svg。
+
+    其中图形svg取所有路径点的矩形范围，文字svg取文本框范围。
+
+    其中心为布局中心，也是组件坐标系的原点。
