+++ docs/concept/shape_svg.md
@@ -95,7 +95,7 @@
 
 - 路径点列表与关联关键点映射
 
-    存储所有当前的路径点，包括与之相关的关键点列表。当关键点发生改变时，该路径点也发生变化。规则如下：
+    存储所有当前的路径点，包括与之相关的关键点关系。当关键点发生改变时，该路径点也发生变化。规则如下：
 
     - 某个相关关键点被移动：更新该路径点坐标
     - 某个相关关键点被删除：移除该路径点及其相关的规则
@@ -142,6 +142,10 @@
 
         对某个矢量图元素或整个矢量图，设置线条样式、填充样式等。
 
+    - 整体移动
+
+        所有关键点、关键点进行整体移动。
+
 - 绘制
 
     更新结束后进行绘制。
@@ -177,8 +181,8 @@
     每次更新都会进行布局调整。
 
     1. 根据元素，重新规划图形外接矩形
-    2. 更新各个点的组件坐标系坐标
-    3. 更新组件在画布中的位置
+    2. 整体移动
+    3. 更新布局中心
 
 - 更新
